Fix building on Windows (again)
